+++ CHANGELOG.md
@@ -7,7 +7,7 @@
   useful summary for people upgrading their application, not a replication
   of the commit log.
 
-## Unreleased
+## 52.1.0
 
 * Swap inset text top margin with padding ([PR #4634](https://github.com/alphagov/govuk_publishing_components/pull/4634))
 * Allow margin bottom to be configured on the table component ([PR #4638](https://github.com/alphagov/govuk_publishing_components/pull/4638))